requirements.txt
fastapi
uvicorn
sqlalchemy
pytz
python-dotenv
langchain
langchain-google-genai
jinja2
psycopg2
python-multipart
chromadb
fitz
transformers
torch
sentence-transformers
google.generativeai
langchain-google-genai